v1.0 - The First Version
I just realised that I had a 130 commit-long list of an unreadable
fiasco and I decided to squash it. I now understand more the concept
of versionning and I'll try to document my commits more often.

The previous commit is the initial application state that I kept
out of nostalgia, but the rest have been brutally squashed into this one.

The "v1.0" is symbolic, it's a manner to say that the app is
more or less stable and has a decent amount of features.
From now on I'll try to be careful, I promise :)